Change menu item input object to use :decimal instead of :float.

+++ lib/plate_slate_web/schema/menu_types.ex
@@ -47,10 +47,10 @@ defmodule PlateSlateWeb.Schema.MenuTypes do
     field :tag, :string
 
     @desc "Priced above a value"
-    field :priced_above, :float
+    field :priced_above, :decimal
 
     @desc "Priced below a value"
-    field :priced_below, :float
+    field :priced_below, :decimal
 
     @desc "Added to the menu before this date"
     field :added_before, :date
